+++ priv/repo/migrations/20190628115309_drop_business_concept_aliases.exs
@@ -1,0 +1,8 @@
+defmodule TdBg.Repo.Migrations.DropBusinessConceptAliases do
+  use Ecto.Migration
+
+  def change do
+    drop(index(:business_concept_aliases, [:business_concept_id]))
+    drop(table("business_concept_aliases"))
+  end
+end