+++ priv/repo/migrations/20190628121631_alter_business_concept_remove_parent_id.exs
@@ -1,0 +1,7 @@
+defmodule TdBg.Repo.Migrations.AlterBusinessConceptRemoveParentId do
+  use Ecto.Migration
+
+  def change do
+    alter(table(:business_concepts), do: remove(:parent_id))
+  end
+end